+++ VTK/Hybrid/vtkPointWidget.cxx
@@ -266,7 +266,7 @@ int vtkPointWidget::DetermineConstraintAxis(int constraint, double *x)
   // outside the hot spot
   if ( ! this->WaitingForMotion )
     {
-    float p[3], d2, tol;
+    double p[3], d2, tol;
     this->CursorPicker->GetPickPosition(p);
     d2 = vtkMath::Distance2BetweenPoints(p,this->LastPickPosition);
     tol = this->HotSpotSize*this->InitialLength;
